User-defined region proposals
The MaskRCNN class is replaced with 2 classes:
1) MaskRCNNDirected: For detecting the objects based on either the user-defined region proposals or the region proposals generated by the RPN.
2) MaskRCNNDirectedRPN: For returning the outputs of the region proposal layer. It either returns the region proposals generated by the RPN or [for error checking] returning the user-defined region proposals.

The decision of whether the region proposals are user-defined or generated by the RPN is made based on the REGION_PROPOSALS attribute in the mrcnn_directed.config.Config class. If this attribute is None, then the region proposals generated by the RPN are used. Otherwise, the user-defined region proposals are used.

In addition to the ProposalLayer, another layer is supported:
1) ProposalLayerDirected: Used when the user feeds a pre-defined region proposals.

In the mrcnn_directed.config.Config class, there is an attribute called REGION_PROPOSALS. If it is None, then the ProposalLayer is used. Otherwise, the ProposalLayerDirected is used.